+++ src/main/frontend/src/store/Store.js
@@ -1,7 +1,7 @@
 import { createStore, applyMiddleware, compose } from 'redux';
 import createSagaMiddleware from 'redux-saga';
 
-import reducer from './reducers/EmployeesReducer';
+import reducer from './reducers/';
 import mySaga from './sagas/sagas';
 
 const sagaMiddleware = createSagaMiddleware();
