--- src/main/frontend/src/store/actions/actions.js
@@ -1,10 +1,0 @@
-export const REQUEST_API_DATA = 'REQUEST_API_DATA';
-export const RECEIVE_API_DATA = 'RECEIVE_API_DATA';
-
-export const requestApiData = () => ({
-  type: REQUEST_API_DATA
-});
-export const receiveApiData = data => ({
-  type: RECEIVE_API_DATA,
-  data
-});
