+++ src/main/frontend/src/store/reducers/EmployeesReducer.js
@@ -1,9 +1,9 @@
 import { RECEIVE_API_DATA } from '../actions/types';
 
-export default (state = {}, { type, data }) => {
+export default (state = { fetching: true }, { type, data }) => {
   switch (type) {
     case RECEIVE_API_DATA:
-      return data;
+      return { fetching: false, data };
     default:
       return state;
   }
