--- src/main/frontend/src/store/reducers/data.js
@@ -1,10 +1,0 @@
-import { RECEIVE_API_DATA } from '../actions/types';
-
-export default (state = {}, { type, data }) => {
-  switch (type) {
-    case RECEIVE_API_DATA:
-      return data;
-    default:
-      return state;
-  }
-};
